Add support for subscriptions
This commit adds support for subscriptions. The websocket is opened once the query is run and all headers are passed as parameters.

If the ws url or the headers are updated, the connection is closed and a new one will be opened when the query is run.

By default, it uses subscriptions-transport-ws client but an alternate subscriptionsClientBuilder can be provided.

This commit also add debouncing input values for both urls.