Fix graphql dependency issue
The root issue is that graphql-language-service-* packages that are pulled by the current version of codemirror-graphql are listed graphql 0.10.1 as they dependencies...

New versions of these packages exist which are now only listing graphql as peer dependency.

Once a new version of codemirror-graphql is released, it should fix the issue.

For now, I have edited yarn.lock file to remove those dependencies.

Fixes #26